Move to decorator
